Update project structure and add technical analyst
