+++ src/agents/fundamentals.py
@@ -1,0 +1,167 @@
+
+from langchain_core.messages import HumanMessage
+
+from agents.state import AgentState, show_agent_reasoning
+
+import json
+
+##### Fundamental Agent #####
+def fundamentals_agent(state: AgentState):
+    """Analyzes fundamental data and generates trading signals."""
+    show_reasoning = state["metadata"]["show_reasoning"]
+    data = state["data"]
+    metrics = data["financial_metrics"][0]
+    financial_line_item = data["financial_line_items"][0]
+    market_cap = data["market_cap"]
+
+    # Initialize signals list for different fundamental aspects
+    signals = []
+    reasoning = {}
+    
+    # 1. Profitability Analysis
+    profitability_score = 0
+    if metrics["return_on_equity"] > 0.15:  # Strong ROE above 15%
+        profitability_score += 1
+    if metrics["net_margin"] > 0.20:  # Healthy profit margins
+        profitability_score += 1
+    if metrics["operating_margin"] > 0.15:  # Strong operating efficiency
+        profitability_score += 1
+        
+    signals.append('bullish' if profitability_score >= 2 else 'bearish' if profitability_score == 0 else 'neutral')
+    reasoning["Profitability"] = {
+        "signal": signals[0],
+        "details": f"ROE: {metrics['return_on_equity']:.2%}, Net Margin: {metrics['net_margin']:.2%}, Op Margin: {metrics['operating_margin']:.2%}"
+    }
+    
+    # 2. Growth Analysis
+    growth_score = 0
+    if metrics["revenue_growth"] > 0.10:  # 10% revenue growth
+        growth_score += 1
+    if metrics["earnings_growth"] > 0.10:  # 10% earnings growth
+        growth_score += 1
+    if metrics["book_value_growth"] > 0.10:  # 10% book value growth
+        growth_score += 1
+        
+    signals.append('bullish' if growth_score >= 2 else 'bearish' if growth_score == 0 else 'neutral')
+    reasoning["Growth"] = {
+        "signal": signals[1],
+        "details": f"Revenue Growth: {metrics['revenue_growth']:.2%}, Earnings Growth: {metrics['earnings_growth']:.2%}"
+    }
+    
+    # 3. Financial Health
+    health_score = 0
+    if metrics["current_ratio"] > 1.5:  # Strong liquidity
+        health_score += 1
+    if metrics["debt_to_equity"] < 0.5:  # Conservative debt levels
+        health_score += 1
+    if metrics["free_cash_flow_per_share"] > metrics["earnings_per_share"] * 0.8:  # Strong FCF conversion
+        health_score += 1
+        
+    signals.append('bullish' if health_score >= 2 else 'bearish' if health_score == 0 else 'neutral')
+    reasoning["Financial_Health"] = {
+        "signal": signals[2],
+        "details": f"Current Ratio: {metrics['current_ratio']:.2f}, D/E: {metrics['debt_to_equity']:.2f}"
+    }
+    
+    # 4. Price to X ratios
+    pe_ratio = metrics["price_to_earnings_ratio"]
+    pb_ratio = metrics["price_to_book_ratio"]
+    ps_ratio = metrics["price_to_sales_ratio"]
+    
+    price_ratio_score = 0
+    if pe_ratio < 25:  # Reasonable P/E ratio
+        price_ratio_score += 1
+    if pb_ratio < 3:  # Reasonable P/B ratio
+        price_ratio_score += 1
+    if ps_ratio < 5:  # Reasonable P/S ratio
+        price_ratio_score += 1
+        
+    signals.append('bullish' if price_ratio_score >= 2 else 'bearish' if price_ratio_score == 0 else 'neutral')
+    reasoning["Price_Ratios"] = {
+        "signal": signals[3],
+        "details": f"P/E: {pe_ratio:.2f}, P/B: {pb_ratio:.2f}, P/S: {ps_ratio:.2f}"
+    }
+
+    # 5. Calculate intrinsic value and compare to market cap
+    free_cash_flow = financial_line_item.get('free_cash_flow')
+    intrinsic_value = calculate_intrinsic_value(
+        free_cash_flow=free_cash_flow,
+        growth_rate=metrics["earnings_growth"],
+        discount_rate=0.10,
+        terminal_growth_rate=0.03,
+        num_years=5,
+    )
+    if market_cap < intrinsic_value:
+        signals.append('bullish')
+    else:
+        signals.append('bearish')
+
+    reasoning["Intrinsic_Value"] = {
+        "signal": signals[4],
+        "details": f"Intrinsic Value: ${intrinsic_value:,.2f}, Market Cap: ${market_cap:,.2f}"
+    }
+    
+    # Determine overall signal
+    bullish_signals = signals.count('bullish')
+    bearish_signals = signals.count('bearish')
+    
+    if bullish_signals > bearish_signals:
+        overall_signal = 'bullish'
+    elif bearish_signals > bullish_signals:
+        overall_signal = 'bearish'
+    else:
+        overall_signal = 'neutral'
+    
+    # Calculate confidence level
+    total_signals = len(signals)
+    confidence = max(bullish_signals, bearish_signals) / total_signals
+    
+    message_content = {
+        "signal": overall_signal,
+        "confidence": f"{round(confidence * 100)}%",
+        "reasoning": reasoning
+    }
+    
+    # Create the fundamental analysis message
+    message = HumanMessage(
+        content=json.dumps(message_content),
+        name="fundamentals_agent",
+    )
+    
+    # Print the reasoning if the flag is set
+    if show_reasoning:
+        show_agent_reasoning(message_content, "Fundamental Analysis Agent")
+    
+    return {
+        "messages": [message],
+        "data": data,
+    }
+
+def calculate_intrinsic_value(
+    free_cash_flow: float,
+    growth_rate: float = 0.05,
+    discount_rate: float = 0.10,
+    terminal_growth_rate: float = 0.02,
+    num_years: int = 5,
+) -> float:
+    """
+    Computes the discounted cash flow (DCF) for a given company based on the current free cash flow.
+    Use this function to calculate the intrinsic value of a stock.
+    """
+    # Estimate the future cash flows based on the growth rate
+    cash_flows = [free_cash_flow * (1 + growth_rate) ** i for i in range(num_years)]
+
+    # Calculate the present value of projected cash flows
+    present_values = []
+    for i in range(num_years):
+        present_value = cash_flows[i] / (1 + discount_rate) ** (i + 1)
+        present_values.append(present_value)
+
+    # Calculate the terminal value
+    terminal_value = cash_flows[-1] * (1 + terminal_growth_rate) / (discount_rate - terminal_growth_rate)
+    terminal_present_value = terminal_value / (1 + discount_rate) ** num_years
+
+    # Sum up the present values and terminal value
+    dcf_value = sum(present_values) + terminal_present_value
+
+    return dcf_value
