+++ src/agents/sentiment.py
@@ -1,0 +1,59 @@
+
+from langchain_core.messages import HumanMessage
+
+from agents.state import AgentState, show_agent_reasoning
+
+import json
+
+##### Sentiment Agent #####
+def sentiment_agent(state: AgentState):
+    """Analyzes market sentiment and generates trading signals."""
+    data = state["data"]
+    insider_trades = data["insider_trades"]
+    show_reasoning = state["metadata"]["show_reasoning"]
+
+    # Loop through the insider trades, if transaction_shares is negative, then it is a sell, which is bearish, if positive, then it is a buy, which is bullish
+    signals = []
+    for trade in insider_trades:
+        transaction_shares = trade["transaction_shares"]
+        if not transaction_shares:
+            continue
+        if transaction_shares < 0:
+            signals.append("bearish")
+        else:
+            signals.append("bullish")
+
+    # Determine overall signal
+    bullish_signals = signals.count("bullish")
+    bearish_signals = signals.count("bearish")
+    if bullish_signals > bearish_signals:
+        overall_signal = "bullish"
+    elif bearish_signals > bullish_signals:
+        overall_signal = "bearish"
+    else:
+        overall_signal = "neutral"
+
+    # Calculate confidence level based on the proportion of indicators agreeing
+    total_signals = len(signals)
+    confidence = max(bullish_signals, bearish_signals) / total_signals
+
+    message_content = {
+        "signal": overall_signal,
+        "confidence": f"{round(confidence * 100)}%",
+        "reasoning": f"Bullish signals: {bullish_signals}, Bearish signals: {bearish_signals}"
+    }
+
+    # Print the reasoning if the flag is set
+    if show_reasoning:
+        show_agent_reasoning(message_content, "Sentiment Analysis Agent")
+
+    # Create the sentiment message
+    message = HumanMessage(
+        content=json.dumps(message_content),
+        name="sentiment_agent",
+    )
+
+    return {
+        "messages": [message],
+        "data": data,
+    }
