+++ src/agents/state.py
@@ -1,0 +1,51 @@
+from typing import Annotated, Any, Dict, Sequence, TypedDict
+
+import operator
+from langchain_core.messages import BaseMessage
+
+
+import json
+
+
+def merge_dicts(a: Dict[str, Any], b: Dict[str, Any]) -> Dict[str, Any]:
+    return {**a, **b}
+
+# Define agent state
+class AgentState(TypedDict):
+    messages: Annotated[Sequence[BaseMessage], operator.add]
+    data: Annotated[Dict[str, Any], merge_dicts]
+    metadata: Annotated[Dict[str, Any], merge_dicts]
+
+
+
+def show_agent_reasoning(output, agent_name):
+    print(f"\n{'=' * 10} {agent_name.center(28)} {'=' * 10}")
+    
+    def convert_to_serializable(obj):
+        if hasattr(obj, 'to_dict'):  # Handle Pandas Series/DataFrame
+            return obj.to_dict()
+        elif hasattr(obj, '__dict__'):  # Handle custom objects
+            return obj.__dict__
+        elif isinstance(obj, (int, float, bool, str)):
+            return obj
+        elif isinstance(obj, (list, tuple)):
+            return [convert_to_serializable(item) for item in obj]
+        elif isinstance(obj, dict):
+            return {key: convert_to_serializable(value) for key, value in obj.items()}
+        else:
+            return str(obj)  # Fallback to string representation
+    
+    if isinstance(output, (dict, list)):
+        # Convert the output to JSON-serializable format
+        serializable_output = convert_to_serializable(output)
+        print(json.dumps(serializable_output, indent=2))
+    else:
+        try:
+            # Parse the string as JSON and pretty print it
+            parsed_output = json.loads(output)
+            print(json.dumps(parsed_output, indent=2))
+        except json.JSONDecodeError:
+            # Fallback to original string if not valid JSON
+            print(output)
+    
+    print("=" * 48)
