+++ src/tools/api.py
@@ -150,81 +150,3 @@ def get_price_data(
 ) -> pd.DataFrame:
     prices = get_prices(ticker, start_date, end_date)
     return prices_to_df(prices)
-
-def calculate_confidence_level(signals: dict) -> float:
-    """Calculate confidence level based on the difference between SMAs."""
-    sma_diff_prev = abs(signals['sma_5_prev'] - signals['sma_20_prev'])
-    sma_diff_curr = abs(signals['sma_5_curr'] - signals['sma_20_curr'])
-    diff_change = sma_diff_curr - sma_diff_prev
-    # Normalize confidence between 0 and 1
-    confidence = min(max(diff_change / signals['current_price'], 0), 1)
-    return confidence
-
-def calculate_macd(prices_df: pd.DataFrame) -> tuple[pd.Series, pd.Series]:
-    ema_12 = prices_df['close'].ewm(span=12, adjust=False).mean()
-    ema_26 = prices_df['close'].ewm(span=26, adjust=False).mean()
-    macd_line = ema_12 - ema_26
-    signal_line = macd_line.ewm(span=9, adjust=False).mean()
-    return macd_line, signal_line
-
-def calculate_rsi(prices_df: pd.DataFrame, period: int = 14) -> pd.Series:
-    delta = prices_df['close'].diff()
-    gain = (delta.where(delta > 0, 0)).fillna(0)
-    loss = (-delta.where(delta < 0, 0)).fillna(0)
-    avg_gain = gain.rolling(window=period).mean()
-    avg_loss = loss.rolling(window=period).mean()
-    rs = avg_gain / avg_loss
-    rsi = 100 - (100 / (1 + rs))
-    return rsi
-
-def calculate_bollinger_bands(
-    prices_df: pd.DataFrame,
-    window: int = 20
-) -> tuple[pd.Series, pd.Series]:
-    sma = prices_df['close'].rolling(window).mean()
-    std_dev = prices_df['close'].rolling(window).std()
-    upper_band = sma + (std_dev * 2)
-    lower_band = sma - (std_dev * 2)
-    return upper_band, lower_band
-
-
-def calculate_obv(prices_df: pd.DataFrame) -> pd.Series:
-    obv = [0]
-    for i in range(1, len(prices_df)):
-        if prices_df['close'].iloc[i] > prices_df['close'].iloc[i - 1]:
-            obv.append(obv[-1] + prices_df['volume'].iloc[i])
-        elif prices_df['close'].iloc[i] < prices_df['close'].iloc[i - 1]:
-            obv.append(obv[-1] - prices_df['volume'].iloc[i])
-        else:
-            obv.append(obv[-1])
-    prices_df['OBV'] = obv
-    return prices_df['OBV']
-
-def calculate_intrinsic_value(
-    free_cash_flow: float,
-    growth_rate: float = 0.05,
-    discount_rate: float = 0.10,
-    terminal_growth_rate: float = 0.02,
-    num_years: int = 5,
-) -> float:
-    """
-    Computes the discounted cash flow (DCF) for a given company based on the current free cash flow.
-    Use this function to calculate the intrinsic value of a stock.
-    """
-    # Estimate the future cash flows based on the growth rate
-    cash_flows = [free_cash_flow * (1 + growth_rate) ** i for i in range(num_years)]
-
-    # Calculate the present value of projected cash flows
-    present_values = []
-    for i in range(num_years):
-        present_value = cash_flows[i] / (1 + discount_rate) ** (i + 1)
-        present_values.append(present_value)
-
-    # Calculate the terminal value
-    terminal_value = cash_flows[-1] * (1 + terminal_growth_rate) / (discount_rate - terminal_growth_rate)
-    terminal_present_value = terminal_value / (1 + discount_rate) ** num_years
-
-    # Sum up the present values and terminal value
-    dcf_value = sum(present_values) + terminal_present_value
-
-    return dcf_value
